api_posts/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
pydantic-settings==2.9.1
asyncpg==0.30.0
sqlalchemy==2.0.41
pydantic==2.11.5
redis==4.5.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6